Implement multi-page routing
I used react-router to serve different pages based on the URL path. I also added the Home and About pages to src/pages, and tested the functionality. Additionally, active nav states were added, and the styling reflects the page the user is currently on. I will be adding the rest of the pages momentarily.