Add responsiveness to site header
On tablet and mobile views, the band name and desktop nav menu are hidden, and a hamburger menu replaces the nav menu. The header height also decreases to take up less space in the mobile/tablet view. The actual mobile/tablet menu still has to be implemented, but it will slide in from the right and take up the whole screen.