Implement basic functionality of release info modal
Opening and closing the modal is now working. The modal's data is dynamically rendered using the activeRelease state, which is set by the releases array. The onClick handler was passed down from the parent releaseList to each of the releaseItems.

The track lists for multi-song releases still has to be implemented.