Add TrackList component and tune responsiveness
The TrackList component was created, which makes it easy to conditionally render a TrackList using the && operator in the ReleaseInfoModal if it is a multi-song release, and thus has a trackList key-value pair. Since the modal height could become arbitrarily large, the use of scrolling when there is overflow helps keep things looking right.