Fix modal usability issues
If the site address bar is shown on mobile, this prevented the bottom of the modal from being seen,
which meant that users could not click the streaming links for releases. The modal height was reduced to 75% viewport height.

Additionally, there was some wierd focus behavior where if you accidentally scrolled on the background behind the modal,
it was difficult to move focus back to the modal to scroll, potentially leading to users not being able to exit the modal.
On modal open/close, <body>'s overflow value was toggled between 'hidden' and 'visible'