Add release-info-modal fade in
This was a bit challenging to pull off at first. Since originally display: none was set on the container element,
The animations were not happening to begin with. I got around this by setting display: flex, and using scale(0) to effectively
hide the element. Then, when the element was needed, I scaled it back to normal size. This allowed the animation transition to happen
because it had a starting point to move from.